* origin/pr/14541:
  Check for passable tile including obstructing vehicle parts.
  Add map::passable_ter_furn for terrain and furniture only.
  Fix a misleading comment:
  Remove unnecessary parentheses
  Implement `map::passable` and `map::impassable` to replace

* origin/pr/14576:
  float to double + minimalistic astyle

* origin/pr/14577:
  migration of items from melee.json to generic.json

* origin/pr/14578:
  Forbid learning uncraft recipes due to skills at chargen

* origin/pr/14580:
  Make NPCs re-check their inventory on load
  Make NPCs not wake up player with whining about sleep
  Disallow bio weapon giving, remove "didnt use moves" msg

* origin/pr/14582:
  Fix #14556.

* origin/pr/14587:
  Fix wasps being annoying

* origin/pr/14588:
  check for use_tiles when making tile_iso decisions

+++ src/cata_utility.cpp
@@ -283,7 +283,7 @@ double convert_weight( int weight )
     return ret;
 }
 
-double temp_to_celsius( float fahrenheit )
+double temp_to_celsius( double fahrenheit )
 {
     return ( ( fahrenheit - 32.0 ) * 5.0 / 9.0 );
 }
